requirements.txt
argostranslate==1.8.1
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
ctranslate2==2.24.0
Django==4.2.5
djangorestframework==3.14.0
filelock==3.12.4
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.24.0
protobuf==4.24.3
PyPDF2==3.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sentencepiece==0.1.99
sqlparse==0.4.4
stanza==1.1.1
sympy==1.12
torch==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
